# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  http.Statistics:
    properties:
      average_order_value:
        type: number
      orders_delivered:
        type: integer
      orders_pending:
        type: integer
      orders_shipped:
        type: integer
      total_orders:
        type: integer
      total_products:
        type: integer
      total_revenue:
        type: number
      total_users:
        type: integer
    type: object
  http.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample e-commerce API server.
  title: E-commerce API
  version: "1.0"
paths:
...
  /api/v1/admin/export/orders:
    post:
      consumes:
      - application/json
      description: Download all orders as CSV file
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema: {}
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export orders to CSV (Admin only)
      tags:
      - Admin - Export
  /api/v1/admin/export/products:
    post:
      consumes:
      - application/json
      description: Download all products as CSV file
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema: {}
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export products to CSV (Admin only)
      tags:
      - Admin - Export
  /api/v1/admin/export/users:
    post:
      consumes:
      - application/json
      description: Download all users as CSV file
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema: {}
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export users to CSV (Admin only)
      tags:
      - Admin - Export
...
  /api/v1/admin/statistics:
    get:
      consumes:
      - application/json
      description: Retrieve platform statistics including user count, order metrics,
        and revenue
      produces:
      - application/json
      responses:
        "200":
          description: Platform statistics
          schema:
            $ref: '#/definitions/http.Statistics'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get platform statistics (Admin only)
      tags:
      - Admin - Statistics
  /api/v1/admin/users/{id}/block:
    put:
      consumes:
      - application/json
      description: Block a user from accessing the platform
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User blocked successfully
          schema:
            $ref: '#/definitions/http.SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Block user account (Admin only)
      tags:
      - Admin - Users
  /api/v1/admin/users/{id}/unblock:
    put:
      consumes:
      - application/json
      description: Unblock a user and allow access to the platform
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User unblocked successfully
          schema:
            $ref: '#/definitions/http.SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unblock user account (Admin only)
      tags:
      - Admin - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"